fix(lint): Fix or ignore lint errors (#2228)

+++ internal/tests/integration/inputdir_unix_test.go
@@ -33,9 +33,9 @@ func checkFileUlimit(t *testing.T, b uint64) {
 }
 
 func TestInputDir_RespectsUlimit(t *testing.T) {
-	numfiles := 32
+	numfiles := uint32(32)
 	flist := map[string]string{}
-	for i := 0; i < numfiles; i++ {
+	for i := 0; i < int(numfiles); i++ {
 		k := fmt.Sprintf("file_%d", i)
 		flist[k] = fmt.Sprintf("hello world %d\n", i)
 	}
@@ -65,9 +65,9 @@ func TestInputDir_RespectsUlimit(t *testing.T) {
 
 	files, err := os.ReadDir(testdir.Join("out"))
 	assert.NilError(t, err)
-	assert.Equal(t, numfiles, len(files))
+	assert.Equal(t, int(numfiles), len(files))
 
-	for i := 0; i < numfiles; i++ {
+	for i := 0; i < int(numfiles); i++ {
 		f := testdir.Join("out", fmt.Sprintf("file_%d", i))
 		_, err := os.Stat(f)
 		assert.NilError(t, err)
